requirement.txt
streamlit
pillow
pytesseract
rapidfuzz
openai